+++ src/components/dropdown/dropdown.tsx
@@ -4,7 +4,13 @@ import {
     h, Host, Listen, Prop, State, Watch
 } from '@stencil/core';
 import { SelectArrow } from '../select/select-arrow';
-import { WcsButtonMode, WcsButtonShape } from '../button/button-interface';
+import {
+    isWcsButtonSize,
+    WcsButtonMode,
+    WcsButtonShape,
+    WcsButtonSize,
+    WcsButtonSizeValues
+} from '../button/button-interface';
 import { createPopper, Instance } from '@popperjs/core';
 import { WcsDropdownPlacement } from './dropdown-interface';
 import { clickTargetIsElementOrChildren, isEscapeKey, isKeydown, isKeyup } from '../../utils/helpers';
@@ -35,6 +41,9 @@ export class Dropdown implements ComponentInterface {
 
     /** Dropdown's button shape */
     @Prop() shape: WcsButtonShape = 'normal';
+    
+    /** Dropdown's button size */
+    @Prop() size: WcsButtonSize = 'm';
 
     /** Specifies whether the dropdown button is clickable or not */
     @Prop() disabled: boolean = false;
@@ -58,6 +67,11 @@ export class Dropdown implements ComponentInterface {
     }
 
     componentDidLoad() {
+        if (!isWcsButtonSize(this.size)) {
+            console.error(`Invalid size value for wcs-dropdown : "${this.size}". Must be one of "${WcsButtonSizeValues.join(', ')}"`);
+            this.size = "m"; // Default fallback value
+        }
+        
         const wcsButtonElement = this.el.shadowRoot.querySelector('wcs-button');
         const buttonWrapper = wcsButtonElement.shadowRoot.querySelector('button');
         this.buttonTextColor = window.getComputedStyle(buttonWrapper).color;
@@ -162,7 +176,7 @@ export class Dropdown implements ComponentInterface {
     render() {
         return (
             <Host>
-                <wcs-button mode={this.mode} shape={this.shape} disabled={this.disabled}
+                <wcs-button mode={this.mode} shape={this.shape} disabled={this.disabled} size={this.size}
                             onClick={($event) => this.onButtonClick($event)}>
                     <div class="wcs-button-content-wrapper">
                         <slot name="placeholder"/>