+++ stories/components/dropdown/dropdown.stories.ts
@@ -1,6 +1,6 @@
 import { Meta, StoryObj } from '@storybook/web-components';
 import { html } from 'lit-html';
-import { WcsButtonMode, WcsButtonShape } from '../../../src/components/button/button-interface';
+import { WcsButtonMode, WcsButtonShape, WcsButtonSize } from '../../../src/components/button/button-interface';
 // @ts-ignore
 import { WcsDropdownPlacement } from '../../../src/components/dropdown/dropdown-interface';
 // @ts-ignore
@@ -31,6 +31,7 @@ export interface DropdownStoryArgs {
     disabled: boolean,
     mode: WcsButtonMode,
     shape: WcsButtonShape,
+    size: WcsButtonSize,
     placement: WcsDropdownPlacement,
     noArrow: boolean
 }
@@ -40,7 +41,7 @@ export const Default: StoryObj<DropdownStoryArgs> = {
         <div style="height: 250px"></div>  <!-- tag to let the overlay display properly -->
         <div style="display: flex; width: 100%; align-items: center; justify-content: center">
             <wcs-dropdown ?disabled=${args.disabled} mode=${args.mode} shape=${args.shape} placement=${args.placement}
-                          ?no-arrow=${args.noArrow}>
+                          size=${args.size} ?no-arrow=${args.noArrow}>
                 <span slot="placeholder">Dropdown</span>
                 <wcs-dropdown-item>Premier item</wcs-dropdown-item>
                 <wcs-dropdown-header>ACTION HEADER</wcs-dropdown-header>
@@ -78,6 +79,37 @@ export const Disabled: StoryObj<DropdownStoryArgs> = {
         placement: 'bottom-end',
         mode: 'plain',
         shape: 'normal'
+    }
+}
+
+const SizeTemplate = (args: DropdownStoryArgs, size: string) => html`
+    <wcs-dropdown size="${size}"
+                  mode=${args.mode}
+                  shape=${args.shape}
+                  placement=${args.placement}
+                  ?disabled=${args.disabled}
+                  ?no-arrow=${args.noArrow}>
+        <span slot="placeholder">Dropdown ${size.toUpperCase()}</span>
+        <wcs-dropdown-item>Premier item</wcs-dropdown-item>
+        <wcs-dropdown-header>ACTION HEADER</wcs-dropdown-header>
+        <wcs-dropdown-item>Second item test avec un long texte</wcs-dropdown-item>
+        <wcs-dropdown-item>Dernier item</wcs-dropdown-item>
+        <wcs-dropdown-divider></wcs-dropdown-divider>
+        <wcs-dropdown-item>Dernier item</wcs-dropdown-item>
+    </wcs-dropdown>
+`;
+
+
+export const Sizes: StoryObj<DropdownStoryArgs> = {
+    render: (args) => html`
+        <div style="display: flex; height: 500px; gap: var(--wcs-margin); align-items: center; justify-content: center;">
+            ${SizeTemplate(args, 'l')}
+            ${SizeTemplate(args, 'm')}
+            ${SizeTemplate(args, 's')}
+        </div>
+    `,
+    args: {
+        ...Default.args
     }
 }